+++ style_generator/lib/src/annotations/copy_with_config.dart
@@ -3,7 +3,6 @@ import "package:style_generator_annotation/copy_with_generator_annotation.dart";
 import "../builder/config.dart";
 
 class CopyWithConfig extends CopyWith implements Config<CopyWith> {
-
   static const String srcAnnotationName = "CopyWith";
 
   @override