+++ style_generator/lib/src/builder/generator.dart
@@ -8,6 +8,7 @@ import "package:meta/meta.dart";
 
 import "../data/annotated_element.dart";
 import "../data/annotation_converter/annotation_converter.dart";
+import "../data/logger.dart";
 import "../data/lookup_store.dart";
 import "../data/variable.dart";
 import "../extensions/element/class_element_extension.dart";
@@ -98,7 +99,7 @@ abstract base class Generator<A, K, C extends Config<A>> {
   /// use [analyzeFor] to override the analyzing.
   /// This is probably only useful in cases where no constructor is necessary
   /// and only direct access to [AnalyzedClass.fields] is required.
-  Future<AnalyzedClass> analyzeClass(AnnotatedElement<A> annotatedClazz, String? constructorName, {FieldType? analyzeFor}) async {
+  Future<AnalyzedClass> analyzeClass(AnnotatedElement<A> annotatedClazz, String? constructorName, {FieldType? analyzeFor, bool? annotationTypeCheck}) async {
     ClassElement clazz = annotatedClazz.element as ClassElement;
     ConstructorElement? constructor = findConstructor(clazz.constructors, constructorName);
 
@@ -106,7 +107,7 @@ abstract base class Generator<A, K, C extends Config<A>> {
 
     VariableHandler state = VariableHandler(clazz: clazz, constructor: constructor);
     await state.indexConstructorDeclarations(resolver, resolvedLib);
-    if (keyAnnotation != null) state.build(keyAnnotation!, type: analyzeFor);
+    if (keyAnnotation != null) state.build(keyAnnotation!, type: analyzeFor, annotationTypeCheck: annotationTypeCheck);
     state.resolveTypes(resolvedLib, type: analyzeFor);
 
     return AnalyzedClass(clazz: clazz, constructor: constructor, variableHandler: state);