+++ style_generator_annotation/lib/equality_generator_annotation.dart
@@ -2,3 +2,5 @@ library;
 
 export "src/annotations/equality.dart";
 export 'src/annotations/equality_key.dart';
+
+export 'package:collection/collection.dart' show DeepCollectionEquality;